put code into main
